+++ packages/react/src/exampleTypes/Facet.js
@@ -1,14 +1,10 @@
 import React from 'react'
 import _ from 'lodash/fp'
-import * as F from 'futil-js'
 import { observer } from 'mobx-react'
 import { exampleTypes } from 'contexture-client'
 import { Flex } from '../layout/Flex'
 import injectTreeNode from '../utils/injectTreeNode'
-
-// For futil-js
-let toggleElement = (check, val, arr = []) =>
-  (check ? _.pull : F.push)(val, arr)
+import { checkBoxValues } from '../utils/actout'
 
 let CheckboxDefault = props => <input type="checkbox" {...props} />
 let RadioListDefault = ({ value, onChange, options }) => (
@@ -88,26 +84,23 @@ let Facet = injectTreeNode(
         )}
         <SelectAll node={node} tree={tree} Checkbox={Checkbox} />
         {_.map(({ name, count }) => {
-          let checked = _.includes(name, node.values)
-          let toggle = () => {
-            tree.mutate(node.path, {
-              values: toggleElement(checked, name, node.values),
-            })
-          }
+          let lens = tree.lens(node.path, 'values')
           return (
-            <Flex
+            <label
               key={name}
               style={{
                 justifyContent: 'space-between',
                 alignItems: 'baseline',
+                display: 'flex',
+                cursor: 'pointer',
               }}
             >
-              <Checkbox onChange={toggle} checked={checked} />
-              <div style={{ flex: 2, padding: '0 5px' }} onClick={toggle}>
+              <Checkbox {...checkBoxValues(name, lens)} />
+              <div style={{ flex: 2, padding: '0 5px' }}>
                 {display(name) || displayBlank()}
               </div>
               <div>{count}</div>
-            </Flex>
+            </label>
           )
         }, _.get('context.options', node))}
         <Flex style={{ justifyContent: 'space-between', margin: '5px 0' }}>