+++ packages/react/src/layout/LensInput.js
@@ -2,6 +2,6 @@ import React from 'react'
 import { observer } from 'mobx-react'
 import { value } from '../utils/actout'
 
-export default observer(({ lens, Input, placeholder, ...x }) => (
-  <Input {...value(lens)} placeholder={placeholder} {...x} />
+export default observer(({ lens, Input, ...x }) => (
+  <Input {...value(lens)} {...x} />
 ))