+++ packages/react/src/utils/actout.js
@@ -1,32 +1,28 @@
 import * as F from 'futil-js'
 import _ from 'lodash/fp'
+import { includeLens, setsWith } from './futil'
 
-// Act-out for mobx
-export let bind = (field, lens) => e => F.set(e.target[field], lens)
+// Map lens to react event handlers
+let binding = (value, getEventValue) => lens => ({
+  [value]: F.view(lens),
+  onChange: setsWith(getEventValue, lens),
+})
+// React events have relevent fields on `target`
+export let targetBinding = field => binding(field, `target.${field}`)
+
 // Makes binding react input value to observables really easy
-export let value = lens => ({
-  value: F.view(lens),
-  onChange: bind('value', lens),
-})
+export let value = targetBinding('value')
 
-export let checkboxValues = (value, lens) => ({
-  checked: F.view(lens).includes(value),
-  onChange: ({ target: { checked } }) =>
-    F.set(
-      checked ? [value, ...F.view(lens)] : _.without([value], F.view(lens)),
-      lens
-    ),
-})
-
-// Just like checkboxValues, but without destructing event.target.checked
-export let simpleCheckBoxValues = (value, lens) => ({
-  checked: F.view(lens).includes(value),
-  onChange: checked =>
-    F.set(
-      checked ? [...F.view(lens), value] : _.without([value], F.view(lens)),
-      lens
-    ),
-})
+export let checkBoxValues = _.flow(includeLens, targetBinding('checked'))
 
 export let enter = f => e => e.key === 'Enter' && f()
 export let click = f => e => e.stopPropagation() || f(e)
+
+export let hover = lens => ({
+  onMouseOver: F.on(lens),
+  onMouseOut: F.off(lens),
+})
+export let focus = lens => ({
+  onFocus: F.on(lens),
+  onBlur: F.off(lens),
+})