+++ packages/react/src/utils/mobx-react-utils.js
@@ -3,16 +3,6 @@ import * as F from 'futil-js'
 import { observable } from 'mobx'
 import { observer, inject } from 'mobx-react'
 
-// Utils
-export let hover = set => ({
-  onMouseOver: () => F.set(true, set),
-  onMouseOut: () => F.set(false, set),
-})
-export let focus = set => ({
-  onFocus: () => F.set(true, set),
-  onBlur: () => F.set(false, set),
-})
-
 // export let Component = (init, render) => inject(init)(observer(render))
 // export let Component = (init, render) => inject(render ? init : () => {})(observer(render || init))
 export let Component = (init, render, displayName) => {
@@ -23,19 +13,9 @@ export let Component = (init, render, displayName) => {
   return inject(render ? init : () => {})(c)
 }
 
-// standard futil lens wasn't working in inject with mapValues
-export let lensOf = s =>
-  _.reduce(
-    (res, key) => {
-      res[key] = F.lensProp(key, s)
-      return res
-    },
-    {},
-    _.keys(s)
-  )
 export let lenservable = x => {
   let s = observable(x)
-  s.lens = lensOf(s)
+  s.lens = F.lensOf(s)
   return s
 }